chore(xcm): add emulated test for opening HRMP channel
Adds an emulated XCM test that sends a message to the Relay Chain to open an HRMP channel between Frequency and AssetHub. This ensures the channel setup logic is functioning correctly in a simulated environment.
