+++ emulated-test/networks/westend/Cargo.toml
@@ -12,3 +12,4 @@ asset-hub-westend-emulated-chain = { workspace = true }
 emulated-integration-tests-common = { workspace = true }
 westend-emulated-chain = { workspace = true }
 frequency-emulated-chain = { path = "../../chains/frequency-westend" }
+polkadot-runtime-parachains = { workspace = true }
