+++ emulated-test/test/frequency-westend/src/lib.rs
@@ -3,10 +3,7 @@ mod imports {
 	pub use frame_support::{
 		assert_ok,
 		sp_runtime::DispatchResult,
-		traits::{
-			fungible::Inspect,
-			fungibles::{Inspect as FungiblesInspect, Mutate as FungiblesMutate},
-		},
+		traits::{fungible::Inspect, fungibles::Inspect as FungiblesInspect},
 		BoundedVec,
 	};
 
@@ -36,7 +33,8 @@ mod imports {
 				xcm::CheckingAccount as FrequencyCheckingAccount,
 				ExistentialDeposit as FrequencyExistentialDeposit,
 			},
-			FrequencyAssetOwner, FrequencyWestendParaPallet as FrequencyWestendPallet, TreasuryAccount as FrequencyTreasuryAccount,
+			FrequencyAssetOwner, FrequencyWestendParaPallet as FrequencyWestendPallet,
+			TreasuryAccount as FrequencyTreasuryAccount,
 		},
 		westend_emulated_chain::{genesis::ED as WESTEND_ED, WestendRelayPallet as WestendPallet},
 		AssetHubWestendPara as AssetHubWestend,
