+++ emulated-test/test/frequency-westend/src/tests/mod.rs
@@ -10,6 +10,6 @@ mod teleport_xfrqcy_to_assethub_with_dot_fee;
 
 mod teleport_xfrqcy_with_dot_fee_from_assethub;
 
-mod send_xcm_to_relay;
+mod send_open_channel_request;
 
 mod utils;
