+++ emulated-test/test/frequency-westend/src/tests/utils.rs
@@ -51,18 +51,6 @@ pub fn create_frequency_asset_on_ah() {
 }
 
 pub fn mint_dot_on_frequency(
-	account: AccountIdOf<<FrequencyWestend as Chain>::Runtime>,
-	amount: Balance,
-) {
-	FrequencyWestend::execute_with(|| {
-		type ForeignAssets = <FrequencyWestend as FrequencyWestendPallet>::ForeignAssets;
-		let dot_asset_id = Parent.into();
-
-		let _ = <ForeignAssets as FungiblesMutate<_>>::mint_into(dot_asset_id, &account, amount);
-	});
-}
-
-pub fn mint_dot_on_frequency_v2(
 	beneficiary: AccountIdOf<<FrequencyWestend as Chain>::Runtime>,
 	amount_to_mint: Balance,
 ) {
@@ -95,7 +83,6 @@ pub fn mint_dot_on_frequency_v2(
 		);
 	});
 }
-
 
 pub fn mint_xrqcy_on_asset_hub(
 	beneficiary: AccountIdOf<<AssetHubWestend as Chain>::Runtime>,
